+++ CHANGELOG.md
@@ -7,6 +7,7 @@ _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2025-01-12_
 * OpenGL: Improved behavior for applications which have multiple OpenGL instances (#4693) @onur-akaydin
 * Multiplot: Improved interactive behavior in .NET Maui, Blazor, and WinUI controls (#4691) @King-Taz
 * IPlotControl: Refactor to support `IMultiplot` allowing users to inject their own multi-plot logic (#4687)
+* Cookbook: Added a colors page demonstrating all colors included with ScottPlot (#4681) @quantfreedom
 
 ## ScottPlot 5.0.52
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2025-01-08_